- Updated the status change field "option" to an int rather than short. This fixes the option values for xmas/sightblaster not getting set, which could explain why some people crash when equipping weapons while in xmas suit. - Also did the same to opt3 since the new packet version uses LONG to store the opt3 values. - Made the clif weapon view function use the EQI constants instead of ints.
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@9337 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/common/mmo.h
@@ -219,7 +219,7 @@ struct mmo_charstatus {
 	short class_;
 	unsigned short status_point,skill_point;
 	int hp,max_hp,sp,max_sp;
-	unsigned short option;
+	unsigned int option;
 	short manner;
 	unsigned char karma;
 	short hair,hair_color,clothes_color;
